# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Worship Core Backend
  version: 0.1.0
  description: Server-side session authentication; mutations require the CSRF header obtained from /api/auth/csrf.
paths:
...
  /api/workspaces/{workspaceId}/scores:
    get:
      operationId: listScores
      description: Workspace score metadata only
      security:
      - session: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreServiceScoreViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input or file
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Provider unavailable or not configured
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: uploadScore
      description: Multipart file and optional songId; actual format, MIME and size validated
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreServiceScoreViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input or file
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Provider unavailable or not configured
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - workspaceId
              - file
              properties:
                workspaceId:
                  type: integer
                  format: int64
                songId:
                  type: integer
                  format: int64
                file:
                  type: string
                  format: binary
  /api/workspaces/{workspaceId}/scores/{scoreId}/download:
    get:
      operationId: downloadScore
      description: Active workspace membership and tenant-scoped score required
      security:
      - session: []
      responses:
        '200':
          description: Capability completed
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input or file
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Provider unavailable or not configured
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: scoreId
        in: path
        required: true
        schema:
          type: integer
          format: int64
  /api/workspaces/{workspaceId}/references:
    get:
      operationId: listReferences
      description: Workspace reference DTOs
      security:
      - session: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReferenceServiceReferenceViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input or file
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Provider unavailable or not configured
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: registerReferenceUrl
      description: Explicit HTTP(S) url and optional title
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceServiceReferenceViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input or file
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Provider unavailable or not configured
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferenceControllerDirectRequest'
  /api/workspaces/{workspaceId}/references/youtube:
    post:
      operationId: registerYoutubeReference
      description: Explicit videoId and optional title
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceServiceReferenceViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input or file
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Provider unavailable or not configured
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferenceControllerVideoRequest'
  /api/workspaces/{workspaceId}/references/youtube/candidates:
    get:
      operationId: searchYoutubeReferences
      description: Transient public video candidates for query parameter
      security:
      - session: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoSearchVideoResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input or file
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Provider unavailable or not configured
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: query
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: \S
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: SESSION
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: Use actual headerName/token from /api/auth/csrf, not a long-lived cached token
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
...
    ReferenceControllerDirectRequest:
      type: object
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 2048
          pattern: \S
          description: Absolute HTTP(S) URL with host and no userinfo; never fetched by Backend.
        title:
          anyOf:
          - type: string
            minLength: 0
            maxLength: 200
          - type: 'null'
      required:
      - url
    ReferenceControllerVideoRequest:
      type: object
      properties:
        videoId:
          type: string
          minLength: 11
          maxLength: 11
          pattern: ^[A-Za-z0-9_-]{11}$
        title:
          anyOf:
          - type: string
            minLength: 0
            maxLength: 200
          - type: 'null'
      required:
      - videoId
    ReferenceServiceReferenceViewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
        title:
          anyOf:
          - type: string
          - type: 'null'
        videoId:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - id
      - url
      - title
      - videoId
    ScoreServiceScoreViewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        songId:
          anyOf:
          - type: integer
            format: int64
          - type: 'null'
        filename:
          type: string
        mediaType:
          type: string
        byteSize:
          type: integer
          format: int64
      required:
      - id
      - songId
      - filename
      - mediaType
      - byteSize
...
    VideoSearchVideoResponse:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        channel:
          type: string
      required:
      - videoId
      - title
      - channel
...
  responses:
    Error400:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - VALIDATION_FAILED
    Error401:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - AUTHENTICATION_REQUIRED
    Error403:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - ACCESS_DENIED
                  - REAUTHENTICATION_REQUIRED
                  - YOUTUBE_AUTHORIZATION_REQUIRED
    Error404:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - NOT_FOUND
                  - REQUEST_REJECTED
    Error405:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error406:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error409:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - STATE_CONFLICT
                  - VERSION_CONFLICT
                  - CONSTRAINT_CONFLICT
    Error415:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error500:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - INTERNAL_ERROR
    Error503:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - PROVIDER_NOT_CONFIGURED
                  - CREDENTIAL_KEY_NOT_CONFIGURED
                  - EXTERNAL_PROVIDER_FAILURE
                  - OBJECT_STORAGE_FAILURE
                  - EXPORT_INTEGRITY_FAILURE